bot/requirements.txt
aiohttp
azure-ai-language-conversations
azure-core-exceptions
azure-identity
azure-keyvault-secrets
botbuilder-core
botbuilder-integration-aiohttp
flask
pyjwt
pytest
pytest-asyncio
python-dotenv
requests
